Updating to remove gonze_lee_nac.py file
